Move some of the methods in the efit updater to the efit utils file
