+++ zero/gkyl_efit_priv.h
@@ -3,3 +3,6 @@
 bool 
 newton_raphson(const double *coeffs, double *xsol);
 
+void
+find_xpts(gkyl_efit* up);
+
